pyproject.toml
[tool.poetry]
name = "reproductionairbnb"
version = "0.1.0"
description = ""
authors = ["procendp <procendp@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.3"
pillow = "^11.0.0"
djangorestframework = "^3.15.2"
pyjwt = "^2.10.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
